+++ components/core/include/memfault/core/platform/debug_log.h
@@ -20,21 +20,23 @@ extern "C" {
 #endif
 
 typedef enum {
-  kMemfaultPlatformLogLevel_Debug,
+  kMemfaultPlatformLogLevel_Debug = 0,
   kMemfaultPlatformLogLevel_Info,
   kMemfaultPlatformLogLevel_Warning,
   kMemfaultPlatformLogLevel_Error,
+  // Convenience definition to get the number of possible levels
+  kMemfaultPlatformLogLevel_NumLevels,
 } eMemfaultPlatformLogLevel;
 
 //! Routine for displaying (or capturing) a log.
 //!
 //! @note it's expected that the implementation will terminate the log with a newline
-//! @note Even if there is no UART or RTT Console, it's worth considering adding an logging
+//! @note Even if there is no UART or RTT Console, it's worth considering adding a logging
 //! implementation that writes to RAM or flash which allows for post-mortem analysis
 MEMFAULT_PRINTF_LIKE_FUNC(2, 3)
 void memfault_platform_log(eMemfaultPlatformLogLevel level, const char *fmt, ...);
 
-//! Routine for printing a log line as-is, only appending a newlines, but without suffixing or
+//! Routine for printing a log line as-is, only appending a newline, but without suffixing or
 //! appending timestamps, log level info, etc. This is used for debug console-commands where
 //! it would degrade the developer experience when the string would be suffixed/post-fixed with
 //! other characters.