+++ components/metrics/include/memfault/metrics/ids_impl.h
@@ -19,6 +19,7 @@ extern "C" {
 //! Generate extern const char * declarations for all IDs (used in key names):
 #define MEMFAULT_METRICS_KEY_DEFINE(key_name, value_type) \
 extern const char * const g_memfault_metrics_id_##key_name;
+#include "memfault/metrics/heartbeat_config.def"
 #include MEMFAULT_METRICS_USER_HEARTBEAT_DEFS_FILE
 #undef MEMFAULT_METRICS_KEY_DEFINE