+++ tests/src/test_memfault_root_cert.cpp
@@ -15,6 +15,7 @@ extern "C" {
   #include <string.h>
 
   #include "memfault/http/root_certs.h"
+  #include "memfault/util/base64.h"
 }
 
 TEST_GROUP(MemfaultRootCert){
@@ -22,37 +23,20 @@ TEST_GROUP(MemfaultRootCert){
   void teardown() { }
 };
 
-static char prv_get_char_from_word(uint32_t word, int offset) {
-  const char base64_table[] =
-      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-  const uint8_t base64_mask = 0x3f; // one char per 6 bits
-  return base64_table[(word >> (offset * 6)) & base64_mask];
-}
-
 static char *prv_der_to_pem(const uint8_t *der, size_t der_len) {
   const char *cert_header = "-----BEGIN CERTIFICATE-----\n";
   const char *cert_footer = "-----END CERTIFICATE-----\n";
   const size_t header_footer_len = strlen(cert_header) + strlen(cert_footer) + 1;
 
   // each 3 byte binary block is encoded as 4 base64 characters
-  const uint32_t pem_len = 4 * ((der_len + 2) / 3) + header_footer_len;
+  const size_t encode_len = MEMFAULT_BASE64_ENCODE_LEN(der_len);
+  const uint32_t pem_len = encode_len + header_footer_len;
   char *pem = (char *)calloc(1, pem_len);
   CHECK(pem != NULL);
 
   int pem_idx = sprintf(pem, "%s", cert_header);
-
-  for (size_t der_idx = 0; der_idx < der_len;  der_idx += 3) {
-    const uint32_t byte0 = der[der_idx];
-    const uint32_t byte1 = ((der_idx + 1) < der_len) ? der[der_idx + 1] : 0;
-    const uint32_t byte2 = ((der_idx + 2) < der_len) ? der[der_idx + 2] : 0;
-    const uint32_t triple = (byte0 << 16) + (byte1 << 8) + byte2;
-
-    pem[pem_idx++] = prv_get_char_from_word(triple, 3);
-    pem[pem_idx++] = prv_get_char_from_word(triple, 2);
-    pem[pem_idx++] = ((der_idx + 1) < der_len) ? prv_get_char_from_word(triple, 1) : '=';
-    pem[pem_idx++] = ((der_idx + 2) < der_len) ? prv_get_char_from_word(triple, 0) : '=';
-  }
-
+  memfault_base64_encode(der, der_len, &pem[pem_idx]);
+  pem_idx += encode_len;
   sprintf(&pem[pem_idx], "%s", cert_footer);
   return pem;
 }